Add GLink
