+++ RR_Godot/src/Core/Core.csproj
@@ -83,6 +83,7 @@
     <Compile Include="Common\Global.cs" />
     <Compile Include="Urdf\UrdfImporter.cs" />
     <Compile Include="Urdf\UrdfNode.cs" />
+    <Compile Include="Urdf\GLink.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="Plugins\Plugins.csproj">
